Update 1_example_neutron_flux_and_units.ipynb
Hi, I noticed that where 'energy_for_each_reaction' is calculated there is an error. 17.5e6 eV is converted to J, to do this you need to multiply by 1.60218e-19 J/eV. In the script instead there is multiplication by 2.8e-18 (which is the result of the previous multiplication). I hope I have been helpful.
Sorry for the trouble. Have a nice day.
Davide